+++ .gitignore
@@ -9,6 +9,3 @@ fastcrypto-derive/Cargo.lock
 .DS_Store
 
 .idea
-/fastcrypto-tbls/src/weight-reduction/target
-/fastcrypto-tbls/src/weight-reduction/archive
-fastcrypto-tbls/src/weight-reduction/Cargo.lock
