+++ fastcrypto-tbls/Cargo.toml
@@ -10,6 +10,7 @@ repository = "https://github.com/MystenLabs/fastcrypto"
 
 [dependencies]
 fastcrypto = { path = "../fastcrypto", features = ["aes"] }
+num-rational = { version = "0.4.2", features = ["std"] }
 
 rand.workspace = true
 serde.workspace = true
@@ -25,17 +26,10 @@ hex = "0.4.3"
 tap = { version = "1.0.1", features = [] }
 serde-big-array = "0.5.1"
 
-[dependencies.solver]
-path = "src/weight-reduction"
-
-[dependencies.num-rational]
-version = "0.4.2"
-features = ["std"]
-default-features = false
-
 [dev-dependencies]
 criterion = "0.5.1"
 generic-tests = "0.1.2"
+test-case = "3.3.1"
 
 [[bench]]
 name = "polynomial"
