--- fastcrypto-tbls/src/weight-reduction/Cargo.toml
@@ -1,43 +1,0 @@
-[package]
-name = "solver"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-path = "super_swiper.rs"
-
-[dependencies]
-num-rational = { version = "0.4.2", features = ["std"], default-features = false }
-
-[dev-dependencies]
-test-case = "3.3.1"
-
-[profile.release]
-overflow-checks = true
-# Enable to use perf.
-#debug = 1
-
-[profile.bench]
-overflow-checks = true
-
-[lints.rust]
-missing-copy-implementations = "warn"
-missing-debug-implementations = "warn"
-non-ascii-idents = "forbid"
-redundant-imports = "warn"
-redundant-lifetimes = "warn"
-unit-bindings = "warn"
-unnameable-types = "warn"
-unsafe_code = "warn"
-unused-import-braces = "warn"
-unused-lifetimes = "warn"
-unused-qualifications = "warn"
-
-[lints.clippy]
-pedantic = { level = "warn", priority = -1 }
-cast_possible_truncation = "allow"
-items_after_statements = "allow"
-manual_midpoint = "allow"
-missing-errors-doc = "allow"
-missing_panics_doc = "allow"
-must_use_candidate = "allow"
