+++ fastcrypto-tbls/src/weight_reduction/data/sui_real_all.dat
