+++ fastcrypto-tbls/src/weight_reduction/scripts/fetch_all_sui_validators.py
